UI
UI update, added a little to the options screen (seen in UIGallery scene) and made a pop-up for the quit button (if we want to use it.) The hp bar for bosses is also laid out in the gallery if we need it. Also tried to go back and anchor stuff that wasn't already anchored.