Wall Jumping & Tweaks
The player can now wall jump. Move towards a wall to slide down it and press space to leap away.
Modified the player's normal jump so it lasts as long as the player holds down space. Tap space for a short hop and hold it to do a full jump.
The player also now has a few more variables. They can be froze (i.e. can't move) and their double jump can be locked.