Pause Menu partially set up
Pause menu set up. Options screen is accessible as well but the options themselves don't do anything yet. For now u can press ESC to back out of both the options and pause menu's. I have it set up on both the dev_cameracontrols scene and the dev_playercontroller scene currently. Also when using the restart button, unity is doing that weird lighting thing again when reloading the scene. We'll have to figure that out later