One-Way platform fixes
Alright, I've fixed both issues now. The platform now has two child objects, both are only box triggers of the same size. One is for holding the player onto the platform as a child, and the other is for handling the platform movement. The second one is disabled once activated once, and re-enabled when reset. The issue with my first solution was that once the player was parented to the platform and the trigger was disabled, the player would be attached to the platform forever since they can't activate the OnTriggerExit for that script. So I had to create two separate triggers on two separate objects. And both of these are on the default layer, so the player cannot jump off of them. Only the actual platform itself.