Wall jump bug fix + Level 2 (WIP)
Level 2 is still a work in progress. I had to spend some time fixing a bug with the wall jump that would cause the jump input to not get registered by moving a portion of the code outside of the OnControllerColliderHit into the Update function.
If you've been having trouble with wall jumps (particularly right walls for some reason), this should help with that.